Implemented Authentication & AuthGate
Added both user sign in and registration functionality with the "display" name being created as the "fullName" upon registration. The AuthGate will be run upon startup and listen to the stream to determine whether the user is logged in or not. If so, present the user with the "ChatPage", if not, present the user with the "IntroPage".