Fixed issue with UI not updating
Upon authentication, the authgate is supposed to listen to state changes and redirect the user accordingly. However, due to the login/registration pages being too far down in the navigation stack it wasn't work. This was resolved by pushing  the navigation stack back . Fixed some additional issues with how this worked as well.