Removed Navigator.push for Authentication
The original functionality pushed either the user login page or registration page to the navigation stack each time the user clicked on the "Sign Up" or "Sign In" link from each page. This means that each time the user does this, the navigation stack will keep getting bigger. Added a "LoginOrRegister" stateful widget to solve this problem.