Created loading animation & list view scroll
Removed "CircularProgressIndicator" from bottom of screen and added it within the list of chat provider messsages so it displays underneath the most recent message. Additionally,  the page will scroll down to the most recent message sent (including ai chatbot messages) when the user sends a prompt to the AI chatbot.