Added chatbot history for individual user accounts
Conversations are now stored within a 'Messages' document in Firestore for each individual user. Conversations are now loaded into the chat view container from Firestore when logging in and removed when signing out. Each conversation is unique to the user account and will only display for them. Additionally, the chatbot will use the context provided from Firestore to remember the conversation for ease of use.