Implemented automatic scroll on chat page
Was facing scrolling issues upon loading a chunk of messages from Firestore due the context not initialising in time. This was addressed by adding another scrolling function with a delay if _isInitialLoad is true. It's a bit janky sometimes based on loading times but it works well enough for now.