Attempted file upload functionality
Got as far as generating a signed URL from Flask and uploading it to an S3 Bucket on AWS. Can't seem to retrieve the file correctly to process from S3 however, will continue to work on this. Additionally, functionality for this is restricted to Android but may be able to program a workaround as well.