Figured out file upload functionality + web
Was able to get files uploaded to a temporary S3 bucket  (will have to talk to Datu to get access to his one) using a signed URL from Flask and retrieving the contents of the file from the S3 bucket. I realised that this has only been programmed for files with the ".docx" extension which is why it wasn't working. I also added web support through a CORS configuration on AWS for testing purposes.