Implemented file upload & summary analysis
The user can now attach a file to their prompt and upload it, the message displaying with appropriate styling based on the enum value. A method will generate a signed url and upload the document to S3 using that URL and then download the document for processing in which the chatbot will spit back a summary. Currently, only files with the ".docx" extension can be summarised otherwise an error is thrown and the user prompt is not passed with the fille contents. This may or may not be addressed as this was not originally in our proposal features.