Fixed bug with chat history and new users
If a user registers, they won't have a conversation ID assigned until they start a conversation. This meant that since I was attemtping to create a temp history item before saving the messages to Firebase, the conversation ID would always be null and the history item data would never load. Solved this problem with a "determineConversationId" method that will always assign a conversation ID when it is null.