Attempted bug fixing for unhandled LLM load
If the user performs an action that destructs the current of the chat view they were previously on, unexpected behaviour or errors will occur. Attempted to address this by aborting LLM response display if the previous conversation ID is no longer equal to the current but this poses some other problems with conversation title generation.