Implemented Update Email Functionality
This was a little more complicated to test as Firebase requires that a verification email is sent to the new email address upon updating. Therefore, I can't add too much notification in the UI other than a snackbar communicating that an email was sent. Will need to come back to this for proper error handling.