Implemented light mode/dark mode toggle
Additionally, the app theme is initially dependant on the system theme before being able to be overridden in the settings menu. This took a while as I had to replace every fixed colour across the app with a dynamic "ThemeData". Additionally, I had to add another provider class to handle manual changes to the app theme. Additional refactoring includes adding a seperate function for displaying a snack bar, reducing repetitive code and some other general edits and commenting.